+++ src/dsp/mod.rs
@@ -1,3 +1,6 @@
+//! Digital signal processing functions and user interface logic.
+
 pub mod buffer;
+pub mod filters;
 
 pub use crate::dsp::buffer::Samples;