+++ src/util.rs
@@ -1,11 +1,18 @@
+//! Utilites for Sampitor library and tests.
+
 #[cfg(test)]
 pub mod test {
     use crate::dsp::Samples;
-    use crate::io::path;
+    use crate::io::audio;
+    use crate::view::View;
+    use crossterm::event::KeyEvent;
     use std::fmt::Write;
     use std::path::PathBuf;
     use tempfile::NamedTempFile;
+    use tui::backend::Backend;
     use tui::buffer::Buffer;
+    use tui::layout::Rect;
+    use tui::terminal::Frame;
     use unicode_width::UnicodeWidthStr;
 
     /// Returns a string representation of the given buffer for debugging purpose.
@@ -42,7 +49,16 @@ pub mod test {
 
     pub fn temp_wave_file(samples: &Samples) -> eyre::Result<PathBuf> {
         let path = NamedTempFile::new().unwrap().path().to_owned();
-        path::write_samples(&path, samples)?;
+        audio::write_samples(&path, samples)?;
         Ok(path)
+    }
+
+    #[derive(Clone, Debug, Default)]
+    pub struct MockView {}
+
+    impl<B: Backend> View<B> for MockView {
+        fn key_event(&mut self, _event: KeyEvent) {}
+        fn process(&mut self, _samples: &mut Samples) {}
+        fn render<'b>(&mut self, _frame: &mut Frame<'b, B>, _area: Rect) {}
     }
 }