+++ src/view/mod.rs
@@ -1,11 +1,11 @@
+//! User interface components.
+
 mod base;
 pub mod chart;
 pub mod file;
-pub mod filters;
-pub mod menu;
+pub mod filter;
 
 pub use base::View;
-pub use chart::Signal;
+pub use chart::Chart;
 pub use file::File;
-pub use filters::Filters;
-pub use menu::Menu;
+pub use filter::Filters;