+++ tests/cli.rs
@@ -6,14 +6,6 @@ use std::process::Command;
 fn missing_file_error() {
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
 
-    let actual = cmd.arg("this_file_does_not_exist.wav").assert();
+    let actual = cmd.args(&["-f", "this_file_does_not_exist.wav"]).assert();
     actual.failure().code(1);
-}
-
-#[test]
-fn no_args_error() {
-    let expected = predicates::str::contains("The following required arguments were not provided:");
-
-    let actual = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap().assert();
-    actual.failure().code(2).stderr(expected);
 }